style(chsarp): in csharp use dotnet default 4 white space
