+++ packages/csharp/Pleisto.Flappy.Test/LLM/BaichuanTest.cs
@@ -3,16 +3,16 @@ using Pleisto.Flappy.LLM;
 
 namespace Pleisto.Flappy.Tests.LLM
 {
-  public class BaichuanTest : LLMTest<Baichuan>
-  {
-    private string BaichuanToken => Environment.GetEnvironmentVariable("BAICHUAN_API_KEY");
-    private string BaichuanSecret => Environment.GetEnvironmentVariable("BAICHUAN_API_SECRET");
-
-    protected override Baichuan OnLLMCreate()
+    public class BaichuanTest : LLMTest<Baichuan>
     {
-      if (BaichuanToken == null || BaichuanSecret == null)
-        Assert.Pass("SKIP!");
-      return new Baichuan(null, BaichuanToken, BaichuanSecret);
+        private string BaichuanToken => Environment.GetEnvironmentVariable("BAICHUAN_API_KEY");
+        private string BaichuanSecret => Environment.GetEnvironmentVariable("BAICHUAN_API_SECRET");
+
+        protected override Baichuan OnLLMCreate()
+        {
+            if (BaichuanToken == null || BaichuanSecret == null)
+                Assert.Pass("SKIP!");
+            return new Baichuan(null, BaichuanToken, BaichuanSecret);
+        }
     }
-  }
 }
