+++ packages/csharp/Pleisto.Flappy/Exceptions/LLMNotSuccessException.cs
@@ -1,13 +1,13 @@
 namespace Pleisto.Flappy.Exceptions
 {
-  /// <summary>
-  /// LLM not success
-  /// </summary>
-  public class LLMNotSuccessException : Exception
-  {
-    internal LLMNotSuccessException()
-       : base("LLM operation return not success")
+    /// <summary>
+    /// LLM not success
+    /// </summary>
+    public class LLMNotSuccessException : Exception
     {
+        internal LLMNotSuccessException()
+           : base("LLM operation return not success")
+        {
+        }
     }
-  }
 }
