+++ packages/csharp/Pleisto.Flappy/Exceptions/TooMoreRetryException.cs
@@ -1,19 +1,19 @@
 namespace Pleisto.Flappy.Exceptions
 {
-  /// <summary>
-  /// Codeinterpreter too more retry exception
-  /// </summary>
-  public class TooMoreRetryException : Exception
-  {
-    internal TooMoreRetryException(int count, Exception inner)
-      : base($"retried count={count}", inner)
-    {
-      this.RetryCount = count;
-    }
-
     /// <summary>
-    /// Retried count
+    /// Codeinterpreter too more retry exception
     /// </summary>
-    public int RetryCount { get; private set; }
-  }
+    public class TooMoreRetryException : Exception
+    {
+        internal TooMoreRetryException(int count, Exception inner)
+          : base($"retried count={count}", inner)
+        {
+            this.RetryCount = count;
+        }
+
+        /// <summary>
+        /// Retried count
+        /// </summary>
+        public int RetryCount { get; private set; }
+    }
 }
