+++ packages/csharp/Pleisto.Flappy/Features/CodeInterpreter/CodeInterpreterInput.cs
@@ -1,13 +1,13 @@
 namespace Pleisto.Flappy.Features.CodeInterpreter
 {
-  /// <summary>
-  /// Input of CodeInterpreter Feature
-  /// </summary>
-  public class CodeInterpreterInput
-  {
     /// <summary>
-    /// Call code
+    /// Input of CodeInterpreter Feature
     /// </summary>
-    public string Code { get; set; }
-  }
+    public class CodeInterpreterInput
+    {
+        /// <summary>
+        /// Call code
+        /// </summary>
+        public string Code { get; set; }
+    }
 }
