+++ packages/csharp/Pleisto.Flappy/Features/CodeInterpreter/CodeInterpreterOutput.cs
@@ -1,13 +1,13 @@
 namespace Pleisto.Flappy.Features.CodeInterpreter
 {
-  /// <summary>
-  /// Result of CodeInterpreter Feature
-  /// </summary>
-  public class CodeInterpreterOutput
-  {
     /// <summary>
-    /// Execute Result
+    /// Result of CodeInterpreter Feature
     /// </summary>
-    public string Result { get; set; }
-  }
+    public class CodeInterpreterOutput
+    {
+        /// <summary>
+        /// Execute Result
+        /// </summary>
+        public string Result { get; set; }
+    }
 }
