+++ packages/csharp/Pleisto.Flappy/Features/FeaturesDefinitionBase.cs
@@ -2,35 +2,35 @@ using Newtonsoft.Json;
 
 namespace Pleisto.Flappy.Features
 {
-  /// <summary>
-  /// Base Definition of Features
-  /// </summary>
-  /// <typeparam name="TArgs">Features call Args</typeparam>
-  /// <typeparam name="TReturns">Features Return</typeparam>
-  public class FeaturesDefinitionBase<TArgs, TReturns>
-    where TArgs : class
-    where TReturns : class
-  {
     /// <summary>
-    /// Feature Name
+    /// Base Definition of Features
     /// </summary>
-    public string Name { get; set; }
+    /// <typeparam name="TArgs">Features call Args</typeparam>
+    /// <typeparam name="TReturns">Features Return</typeparam>
+    public class FeaturesDefinitionBase<TArgs, TReturns>
+      where TArgs : class
+      where TReturns : class
+    {
+        /// <summary>
+        /// Feature Name
+        /// </summary>
+        public string Name { get; set; }
 
-    /// <summary>
-    /// Feature Description
-    /// </summary>
-    public string Description { get; set; } = null;
+        /// <summary>
+        /// Feature Description
+        /// </summary>
+        public string Description { get; set; } = null;
 
-    /// <summary>
-    /// Feature Argument
-    /// </summary>
-    [JsonRequired]
-    public TArgs Args { get; set; }
+        /// <summary>
+        /// Feature Argument
+        /// </summary>
+        [JsonRequired]
+        public TArgs Args { get; set; }
 
-    /// <summary>
-    /// Feature ReturnType
-    /// </summary>
-    [JsonRequired]
-    public TReturns ReturnType { get; set; }
-  }
+        /// <summary>
+        /// Feature ReturnType
+        /// </summary>
+        [JsonRequired]
+        public TReturns ReturnType { get; set; }
+    }
 }
