+++ packages/csharp/Pleisto.Flappy/Features/Synthesized/SynthesizedFeature.cs
@@ -6,31 +6,31 @@ using Pleisto.Flappy.Utils;
 
 namespace Pleisto.Flappy.Features.Syntehesized
 {
-  /// <summary>
-  /// Synthesized Feature
-  /// </summary>
-  /// <typeparam name="TArgs">Argument of feature</typeparam>
-  /// <typeparam name="TReturn">Return of feature</typeparam>
-  /// <typeparam name="TOptions">Options of feature</typeparam>
-  public class SynthesizedFeature<TArgs, TReturn, TOptions> : FlappyFeatureBase<TArgs, TReturn, TOptions>, IFlappyFeature
-    where TArgs : class
-    where TReturn : class
-    where TOptions : FlappyFeatureOption
-  {
     /// <summary>
-    /// Feature Call
+    /// Synthesized Feature
     /// </summary>
-    /// <param name="agent"></param>
-    /// <param name="args"></param>
-    /// <returns></returns>
-    public override async Task<TReturn> Call(FlappyAgent agent, TArgs args)
+    /// <typeparam name="TArgs">Argument of feature</typeparam>
+    /// <typeparam name="TReturn">Return of feature</typeparam>
+    /// <typeparam name="TOptions">Options of feature</typeparam>
+    public class SynthesizedFeature<TArgs, TReturn, TOptions> : FlappyFeatureBase<TArgs, TReturn, TOptions>, IFlappyFeature
+      where TArgs : class
+      where TReturn : class
+      where TOptions : FlappyFeatureOption
     {
-      var describe = Define.Description;
-      var schema = FlappyAgent.GetSchemaGenerator();
-      var returnTypeSchema = schema.Generate(typeof(TReturn)); //extractSchema(callingSchema, "returnType");
-      var argsSchema = schema.Generate(typeof(TArgs)); ;
-      var prompt = JObject.FromObject(args); //(args as object) is string ? args as string : JObject.FromObject(args).ToString();
-      var originalRequestMessage = new ChatMLMessage[]{
+        /// <summary>
+        /// Feature Call
+        /// </summary>
+        /// <param name="agent"></param>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        public override async Task<TReturn> Call(FlappyAgent agent, TArgs args)
+        {
+            var describe = Define.Description;
+            var schema = FlappyAgent.GetSchemaGenerator();
+            var returnTypeSchema = schema.Generate(typeof(TReturn)); //extractSchema(callingSchema, "returnType");
+            var argsSchema = schema.Generate(typeof(TArgs)); ;
+            var prompt = JObject.FromObject(args); //(args as object) is string ? args as string : JObject.FromObject(args).ToString();
+            var originalRequestMessage = new ChatMLMessage[]{
                   new ChatMLMessage
                   {
                     Role = ChatMLMessageRole.System,
@@ -50,27 +50,27 @@ namespace Pleisto.Flappy.Features.Syntehesized
                     })
                   }
               };
-      var requestMessage = originalRequestMessage;
-      ChatMLResponse result = null;
-      int retry = Options?.Retry ?? agent.retry;
+            var requestMessage = originalRequestMessage;
+            ChatMLResponse result = null;
+            int retry = Options?.Retry ?? agent.retry;
 
-      while (true)
-        try
-        {
-          result = await agent.llm.ChatComplete(requestMessage);
-          var data = ParseComplete(result);
+            while (true)
+                try
+                {
+                    result = await agent.llm.ChatComplete(requestMessage);
+                    var data = ParseComplete(result);
 
-          return data;
-        }
-        catch (Exception ex)
-        {
-          if (retry <= 0)
-            throw new TooMoreRetryException(retry, ex);
-          retry -= 1;
-          if (result?.Success == true && (result.Data?.Length ?? 0) > 0)
-          {
-            requestMessage = originalRequestMessage.Union(new ChatMLMessage[]
-            {
+                    return data;
+                }
+                catch (Exception ex)
+                {
+                    if (retry <= 0)
+                        throw new TooMoreRetryException(retry, ex);
+                    retry -= 1;
+                    if (result?.Success == true && (result.Data?.Length ?? 0) > 0)
+                    {
+                        requestMessage = originalRequestMessage.Union(new ChatMLMessage[]
+                        {
               new ChatMLMessage{
                 Role = ChatMLMessageRole.Assistant,
                 Content = result?.Data ?? ""
@@ -83,36 +83,36 @@ namespace Pleisto.Flappy.Features.Syntehesized
                   ["message"] = ex.Message
                 })
               }
-            }).ToArray();
-          }
+                        }).ToArray();
+                    }
+                }
+        }
+
+        private static TReturn ParseComplete(ChatMLResponse msg)
+        {
+            var startIdx = msg.Data?.IndexOf('{') ?? -1;
+            var endIdx = msg.Data?.LastIndexOf('}') ?? -1;
+            if (startIdx == -1 || endIdx == -1 || endIdx < startIdx)
+                throw new InvalidJsonDataException(startIdx, endIdx, msg.Data);
+
+            var content = msg.Data.Substring(startIdx, endIdx - startIdx + 1).Trim();
+            try
+            {
+                var json = JObject.Parse(content);
+                return json.ToObject<TReturn>();
+            }
+            catch (Exception ex)
+            {
+                throw new InvalidJsonDataException(startIdx, endIdx, msg.Data, content, ex);
+            }
+        }
+
+        /// <summary>
+        /// Create Synthesized Feature
+        /// </summary>
+        /// <param name="define">Synthesized Definition</param>
+        public SynthesizedFeature(SynthesizedFeatureDefinition<TArgs, TReturn> define) : base(define)
+        {
         }
     }
-
-    private static TReturn ParseComplete(ChatMLResponse msg)
-    {
-      var startIdx = msg.Data?.IndexOf('{') ?? -1;
-      var endIdx = msg.Data?.LastIndexOf('}') ?? -1;
-      if (startIdx == -1 || endIdx == -1 || endIdx < startIdx)
-        throw new InvalidJsonDataException(startIdx, endIdx, msg.Data);
-
-      var content = msg.Data.Substring(startIdx, endIdx - startIdx + 1).Trim();
-      try
-      {
-        var json = JObject.Parse(content);
-        return json.ToObject<TReturn>();
-      }
-      catch (Exception ex)
-      {
-        throw new InvalidJsonDataException(startIdx, endIdx, msg.Data, content, ex);
-      }
-    }
-
-    /// <summary>
-    /// Create Synthesized Feature
-    /// </summary>
-    /// <param name="define">Synthesized Definition</param>
-    public SynthesizedFeature(SynthesizedFeatureDefinition<TArgs, TReturn> define) : base(define)
-    {
-    }
-  }
 }
