+++ packages/csharp/Pleisto.Flappy/Interfaces/FlappyFeatureOption.cs
@@ -1,13 +1,13 @@
 namespace Pleisto.Flappy.Interfaces
 {
-  /// <summary>
-  /// Flappy Feature Options
-  /// </summary>
-  public class FlappyFeatureOption
-  {
     /// <summary>
-    /// Retry of Feature Call
+    /// Flappy Feature Options
     /// </summary>
-    public int? Retry { get; set; }
-  }
+    public class FlappyFeatureOption
+    {
+        /// <summary>
+        /// Retry of Feature Call
+        /// </summary>
+        public int? Retry { get; set; }
+    }
 }
