+++ packages/csharp/Pleisto.Flappy/LLM/Interfaces/ChatMLMessage.cs
@@ -1,18 +1,18 @@
 namespace Pleisto.Flappy.LLM.Interfaces
 {
-  /// <summary>
-  /// LLM Message
-  /// </summary>
-  public class ChatMLMessage
-  {
     /// <summary>
-    /// Message Role
+    /// LLM Message
     /// </summary>
-    public ChatMLMessageRole Role { get; set; }
+    public class ChatMLMessage
+    {
+        /// <summary>
+        /// Message Role
+        /// </summary>
+        public ChatMLMessageRole Role { get; set; }
 
-    /// <summary>
-    /// Message Content
-    /// </summary>
-    public string Content { get; set; }
-  }
+        /// <summary>
+        /// Message Content
+        /// </summary>
+        public string Content { get; set; }
+    }
 }
