+++ packages/csharp/Pleisto.Flappy/LLM/Interfaces/ChatMLMessageRole.cs
@@ -1,23 +1,23 @@
 namespace Pleisto.Flappy.LLM.Interfaces
 {
-  /// <summary>
-  /// LLM Message Role
-  /// </summary>
-  public enum ChatMLMessageRole
-  {
     /// <summary>
-    /// System
+    /// LLM Message Role
     /// </summary>
-    System,
+    public enum ChatMLMessageRole
+    {
+        /// <summary>
+        /// System
+        /// </summary>
+        System,
 
-    /// <summary>
-    /// User
-    /// </summary>
-    User,
+        /// <summary>
+        /// User
+        /// </summary>
+        User,
 
-    /// <summary>
-    /// Assistant
-    /// </summary>
-    Assistant,
-  }
+        /// <summary>
+        /// Assistant
+        /// </summary>
+        Assistant,
+    }
 }
