+++ packages/csharp/Pleisto.Flappy/LLM/Interfaces/ChatMLResponse.cs
@@ -1,18 +1,18 @@
 namespace Pleisto.Flappy.LLM.Interfaces
 {
-  /// <summary>
-  /// LLM Response
-  /// </summary>
-  public class ChatMLResponse
-  {
     /// <summary>
-    /// Response is success
+    /// LLM Response
     /// </summary>
-    public bool Success { get; set; }
+    public class ChatMLResponse
+    {
+        /// <summary>
+        /// Response is success
+        /// </summary>
+        public bool Success { get; set; }
 
-    /// <summary>
-    /// Data
-    /// </summary>
-    public string Data { get; set; }
-  }
+        /// <summary>
+        /// Data
+        /// </summary>
+        public string Data { get; set; }
+    }
 }
