+++ packages/csharp/Pleisto.Flappy/LLM/Interfaces/GenerateConfig.cs
@@ -1,23 +1,23 @@
 namespace Pleisto.Flappy.LLM.Interfaces
 {
-  /// <summary>
-  /// LLM Generation Config
-  /// </summary>
-  public class GenerateConfig
-  {
     /// <summary>
-    /// MaxTokens
+    /// LLM Generation Config
     /// </summary>
-    public int MaxTokens { get; set; }
+    public class GenerateConfig
+    {
+        /// <summary>
+        /// MaxTokens
+        /// </summary>
+        public int MaxTokens { get; set; }
 
-    /// <summary>
-    /// Temperature
-    /// </summary>
-    public int Temperature { get; set; }
+        /// <summary>
+        /// Temperature
+        /// </summary>
+        public int Temperature { get; set; }
 
-    /// <summary>
-    /// Top P
-    /// </summary>
-    public int Top_P { get; set; }
-  }
+        /// <summary>
+        /// Top P
+        /// </summary>
+        public int Top_P { get; set; }
+    }
 }
