+++ packages/csharp/Pleisto.Flappy/LLM/LLMBase.cs
@@ -2,22 +2,22 @@ using Pleisto.Flappy.LLM.Interfaces;
 
 namespace Pleisto.Flappy.LLM
 {
-  /// <summary>
-  /// LLM Base
-  /// </summary>
-  public interface ILLMBase
-  {
     /// <summary>
-    /// Max Tokens of LLM
+    /// LLM Base
     /// </summary>
-    int MaxTokens { get; }
+    public interface ILLMBase
+    {
+        /// <summary>
+        /// Max Tokens of LLM
+        /// </summary>
+        int MaxTokens { get; }
 
-    /// <summary>
-    /// Run LLM Complete
-    /// </summary>
-    /// <param name="message">Request message</param>
-    /// <param name="config">Generation Configure</param>
-    /// <returns></returns>
-    Task<ChatMLResponse> ChatComplete(ChatMLMessage[] message, GenerateConfig config = null);
-  }
+        /// <summary>
+        /// Run LLM Complete
+        /// </summary>
+        /// <param name="message">Request message</param>
+        /// <param name="config">Generation Configure</param>
+        /// <returns></returns>
+        Task<ChatMLResponse> ChatComplete(ChatMLMessage[] message, GenerateConfig config = null);
+    }
 }
